+++ src/main/java/com/secuconnect/client/api/GeneralContractsApi.java
@@ -6,7 +6,6 @@ import com.secuconnect.client.model.GeneralContractsList;
 import com.secuconnect.client.model.GeneralContractsProductModel;
 import com.secuconnect.client.model.GetAvailablePaymentMethodsDTO;
 import com.secuconnect.client.model.ProductExceptionPayload;
-import com.secuconnect.client.model.StringList;
 import okhttp3.Call;
 
 import java.lang.reflect.Type;
@@ -193,11 +192,11 @@ public class GeneralContractsApi {
      * Get available payment methods for given contract
      * @param generalContractId Contract identifier (required)
      * @param body options
-     * @return StringList
+     * @return List&lt;String&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public StringList getAvailablePaymentMethods(String generalContractId, GetAvailablePaymentMethodsDTO body) throws ApiException {
-        ApiResponse<StringList> resp = getAvailablePaymentMethodsWithHttpInfo(generalContractId, body);
+    public List<String> getAvailablePaymentMethods(String generalContractId, GetAvailablePaymentMethodsDTO body) throws ApiException {
+        ApiResponse<List<String>> resp = getAvailablePaymentMethodsWithHttpInfo(generalContractId, body);
         return resp.getData();
     }
 
@@ -206,12 +205,12 @@ public class GeneralContractsApi {
      * Get available payment methods for given contract
      * @param generalContractId Contract identifier (required)
      * @param body options
-     * @return ApiResponse&lt;StringList&gt;
+     * @return ApiResponse&lt;List&lt;String&gt;&gt;
      * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the response body
      */
-    public ApiResponse<StringList> getAvailablePaymentMethodsWithHttpInfo(String generalContractId, GetAvailablePaymentMethodsDTO body) throws ApiException {
+    public ApiResponse<List<String>> getAvailablePaymentMethodsWithHttpInfo(String generalContractId, GetAvailablePaymentMethodsDTO body) throws ApiException {
         Call call = getAvailablePaymentMethodsValidateBeforeCall(generalContractId, body);
-        Type localVarReturnType = new TypeToken<StringList>(){}.getType();
+        Type localVarReturnType = new TypeToken<List<String>>(){}.getType();
         return apiClient.execute(call, localVarReturnType);
     }
 
@@ -224,9 +223,9 @@ public class GeneralContractsApi {
      * @return The request call
      * @throws ApiException If fail to process the API call, e.g. serializing the request body object
      */
-    public Call getAvailablePaymentMethodsAsync(String generalContractId, GetAvailablePaymentMethodsDTO body, final ApiCallback<StringList> callback) throws ApiException {
+    public Call getAvailablePaymentMethodsAsync(String generalContractId, GetAvailablePaymentMethodsDTO body, final ApiCallback<List<String>> callback) throws ApiException {
         Call call = getAvailablePaymentMethodsValidateBeforeCall(generalContractId, body);
-        Type localVarReturnType = new TypeToken<StringList>(){}.getType();
+        Type localVarReturnType = new TypeToken<List<String>>(){}.getType();
         apiClient.executeAsync(call, localVarReturnType, callback);
         return call;
     }