+++ src/main/java/com/secuconnect/client/model/SecupayTransactionCaptureDTO.java
@@ -1,6 +1,8 @@
 package com.secuconnect.client.model;
 
 import com.google.gson.annotations.SerializedName;
+import com.secuconnect.client.model.PaymentContractsProductModel;
+import com.secuconnect.client.model.SecupayTransactionSetShippingInformationDTO;
 import java.util.Objects;
 
 /**
@@ -8,23 +10,43 @@ import java.util.Objects;
  */
 public class SecupayTransactionCaptureDTO {
   @SerializedName("contract")
-  private Object contract = null;
+  private PaymentContractsProductModel contract = null;
 
-  public SecupayTransactionCaptureDTO contract(Object contract) {
+  @SerializedName("shipping_information")
+  private SecupayTransactionSetShippingInformationDTO shippingInformation = null;
+
+  public SecupayTransactionCaptureDTO contract(PaymentContractsProductModel contract) {
     this.contract = contract;
     return this;
   }
 
    /**
-   * The payment contract object
+   * Get contract
    * @return contract
   **/
-  public Object getContract() {
+  public PaymentContractsProductModel getContract() {
     return contract;
   }
 
-  public void setContract(Object contract) {
+  public void setContract(PaymentContractsProductModel contract) {
     this.contract = contract;
+  }
+
+  public SecupayTransactionCaptureDTO shippingInformation(SecupayTransactionSetShippingInformationDTO shippingInformation) {
+    this.shippingInformation = shippingInformation;
+    return this;
+  }
+
+   /**
+   * Get shippingInformation
+   * @return shippingInformation
+  **/
+  public SecupayTransactionSetShippingInformationDTO getShippingInformation() {
+    return shippingInformation;
+  }
+
+  public void setShippingInformation(SecupayTransactionSetShippingInformationDTO shippingInformation) {
+    this.shippingInformation = shippingInformation;
   }
 
   @Override
@@ -36,12 +58,13 @@ public class SecupayTransactionCaptureDTO {
       return false;
     }
     SecupayTransactionCaptureDTO secupayTransactionCaptureDTO = (SecupayTransactionCaptureDTO) o;
-    return Objects.equals(this.contract, secupayTransactionCaptureDTO.contract);
+    return Objects.equals(this.contract, secupayTransactionCaptureDTO.contract) &&
+        Objects.equals(this.shippingInformation, secupayTransactionCaptureDTO.shippingInformation);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(contract);
+    return Objects.hash(contract, shippingInformation);
   }
 
   @Override
@@ -49,6 +72,7 @@ public class SecupayTransactionCaptureDTO {
     StringBuilder sb = new StringBuilder();
     sb.append("class SecupayTransactionCaptureDTO {\n");
     sb.append("    contract: ").append(toIndentedString(contract)).append("\n");
+    sb.append("    shippingInformation: ").append(toIndentedString(shippingInformation)).append("\n");
     sb.append("}");
     return sb.toString();
   }